requirements.txt
# Core ML libraries
torch>=1.9.0
sentence-transformers>=2.2.0
transformers>=4.20.0

# Information Retrieval
faiss-cpu>=1.7.0
faiss-gpu>=1.7.0; sys_platform != "darwin"
rank-bm25>=0.2.2
ir-measures>=0.3.3

# Data processing
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web framework for demo
streamlit>=1.20.0
fastapi>=0.85.0
uvicorn>=0.18.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.4.0
